+++ mtTorrent/Core/Uploader.h
@@ -14,7 +14,7 @@ namespace mtt
 		void isInterested(PeerCommunication* p);
 		bool pieceRequest(PeerCommunication* p, PieceBlockInfo& info);
 
-		size_t uploaded = 0;
+		uint64_t uploaded = 0;
 
 	private:
