+++ mtTorrent/utils/BencodeWriter.h
@@ -17,15 +17,15 @@ namespace mtt
 		void startRawArrayItem(const char* text);
 		void endArray();
 	
-		void addNumber(size_t number);
+		void addNumber(uint64_t number);
 		void addText(const char* text);
 		void addText(const std::string& text);
 
-		void addItem(const char* name, size_t number);
+		void addItem(const char* name, uint64_t number);
 		void addItem(const char* name, const char* text);
 		void addItemFromBuffer(const char* name, const char* buffer, size_t size);
 
-		void addRawItem(const char* name, size_t number);
+		void addRawItem(const char* name, uint64_t number);
 		void addRawItem(const char* name, const char* text);
 		void addRawItem(const char* name, const std::string& text);
 		void addRawItemFromBuffer(const char* name, const char* buffer, size_t size);
